+++ packages/admin/src/apis/article/type.ts
@@ -1,0 +1,21 @@
+export type ArticleAddRequest = {
+  title?: string,
+  description?: string,
+  image?: File,
+}
+
+
+export type ArticleDetailResponse = {
+  title: string;
+   description: string;
+   image?: string;
+}
+
+
+export type ArticleManagementEditRequest = {
+  title?: string,
+  description?: string,
+  image?: File,
+  articleId: number
+}
+
