--- packages/admin/src/apis/articleManagement/index.ts
@@ -1,4 +1,0 @@
-export * from './articleManagementAdd'
-export * from './articleManagementDel'
-export * from './articleManagementEdit'
-export * from './articleManagementDetail'
