+++ packages/admin/src/apis/index.ts
@@ -1,3 +1,3 @@
 export * from './login/adminLogin';
 export * from './login/type';
-export * from './articleManagement'
+export * from './article'
